+++ crates/wdl-cli/src/analysis/source.rs
@@ -85,6 +85,16 @@ impl std::str::FromStr for Source {
     }
 }
 
+impl Default for Source {
+    fn default() -> Self {
+        // Default to the current directory.
+        Source::Directory(
+            std::env::current_dir()
+                .unwrap_or_else(|_| PathBuf::from(std::path::Component::CurDir.as_os_str())),
+        )
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;