+++ crates/wdl-cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Changed
+
+* Source now has a default implementation ([#476](https://github.com/stjude-rust-labs/wdl/pull/476)).
+
 ## 0.2.0 - 05-27-2025
 
 #### Dependencies